Revert "Chem 025 sign up"
