+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
 /.build
-/.dynamodb
-.serverless
+# .serverless
