+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.ts",
   "scripts": {
     "build": "tsc -p .",
-    "start:local": "serverless offline start",
+    "start:local": "serverless offline",
     "deploy": "serverless deploy",
     "start": "node ./build/index.js",
     "dev": "cross-env nodemon ./src/index.ts"
@@ -22,7 +22,6 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "joi": "^17.7.0",
-    "lodash": "^4.17.21",
     "serverless-http": "^3.1.0",
     "uuid": "^8.3.2"
   },
@@ -35,7 +34,7 @@
     "concurrently": "^7.6.0",
     "nodemon": "^2.0.20",
     "serverless": "^3.26.0",
-    "serverless-dynamodb-local": "^0.2.35",
+    "serverless-dynamodb-local": "^0.2.40",
     "serverless-offline": "^12.0.3",
     "serverless-plugin-typescript": "^2.1.4",
     "ts-node": "^10.9.1",
