--- src/entities/auth.ts
@@ -1,4 +1,0 @@
-export interface ISignUp {
-    email: string
-    password: string
-}
