--- src/environment/path.ts
@@ -1,3 +1,0 @@
-export const USER_POOL = process.env.USER_POOL;
-export const USER_POOL_CLIENT = process.env.USER_POOL_CLIENT;
-export const DYNAMO = process.env.DYNAMO;
