+++ src/index.ts
@@ -1,7 +1,6 @@
 import express, { Express, NextFunction, Request, Response } from 'express';
 import dotenv from 'dotenv';
-import router from './routers/router';
-import authRouter from './routers/authRouter';
+import router from './router/router';
 import AWS from 'aws-sdk';
 import dayjs from 'dayjs';
 import utc from "dayjs/plugin/utc";
@@ -20,8 +19,6 @@ app.use(cors(corsOptions));
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-
-app.use('/auth', authRouter);
 
 // Route Handlers
 app.use('/api', router);
