--- src/utils/removeWhiteSpace.ts
@@ -1,3 +1,0 @@
-export const removeWhiteSpace = (string: string): string => {
-    return string.replace(/ /g, '');
-}
